requirements.txt
Django==4.2.11
djangorestframework==3.14.0
psycopg[binary]==3.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
